Move src/lib/ to src/util/ for Android
It is because that "lib" conflicts with the prefix automatically
appended to a library name by Android build system.
There may be some solution, but I cannot find it.
