+++ CMakeLists.txt
@@ -90,7 +90,7 @@ if(ANDROID)
   add_library( SDL2_ttf SHARED IMPORTED )
   add_library( SDL2_mixer SHARED IMPORTED )
   add_library( lua SHARED IMPORTED )
-  add_library( lib SHARED IMPORTED )
+  add_library( util SHARED IMPORTED )
   add_library( snail SHARED IMPORTED )
   add_library( elona SHARED IMPORTED )
   add_library( boost_filesystem SHARED IMPORTED )
@@ -107,8 +107,8 @@ if(ANDROID)
     ${DISTRIBUTION_DIR}/${ANDROID_BUILD_DIR}/obj/local/${ANDROID_ABI}/libSDL2_mixer.so)
   set_target_properties(lua PROPERTIES IMPORTED_LOCATION
     ${DISTRIBUTION_DIR}/${ANDROID_BUILD_DIR}/obj/local/${ANDROID_ABI}/liblua.so)
-  set_target_properties(lib PROPERTIES IMPORTED_LOCATION
-    ${DISTRIBUTION_DIR}/${ANDROID_BUILD_DIR}/obj/local/${ANDROID_ABI}/lib.so)
+  set_target_properties(util PROPERTIES IMPORTED_LOCATION
+    ${DISTRIBUTION_DIR}/${ANDROID_BUILD_DIR}/obj/local/${ANDROID_ABI}/libutil.so)
   set_target_properties(snail PROPERTIES IMPORTED_LOCATION
     ${DISTRIBUTION_DIR}/${ANDROID_BUILD_DIR}/obj/local/${ANDROID_ABI}/libsnail.so)
   set_target_properties(elona PROPERTIES IMPORTED_LOCATION
@@ -126,7 +126,7 @@ if(ANDROID)
   include_directories(${EXTERNAL_LIB_LOCATION}/SDL2_ttf)
   include_directories(${EXTERNAL_LIB_LOCATION}/SDL2_mixer)
   include_directories(${EXTERNAL_LIB_LOCATION}/lua/src)
-  include_directories(${EXTERNAL_LIB_LOCATION}/lib)
+  include_directories(${EXTERNAL_LIB_LOCATION}/util)
   include_directories(${EXTERNAL_LIB_LOCATION}/snail)
   include_directories(${EXTERNAL_LIB_LOCATION}/elona)
   include_directories(${EXTERNAL_LIB_LOCATION}/boost/include)
@@ -134,7 +134,7 @@ if(ANDROID)
 
   add_library( ${PROJECT_NAME} SHARED src/main.cpp )
 
-  target_link_libraries( ${PROJECT_NAME} android log SDL2 SDL2_image SDL2_ttf SDL2_mixer lua boost_filesystem boost_system boost_iostreams lib snail elona)
+  target_link_libraries( ${PROJECT_NAME} android log SDL2 SDL2_image SDL2_ttf SDL2_mixer lua boost_filesystem boost_system boost_iostreams util snail elona)
 else()
   # Options
   option(WITH_TESTS "Build test executable (TESTS or BENCH)" "OFF")
@@ -403,9 +403,9 @@ else()
     add_definitions("-D${opt}")
   endforeach()
 
-  # My libraries: lib
-  add_subdirectory(src/lib)
-  target_link_libraries(${PROJECT_NAME} lib)
+  # My libraries: util
+  add_subdirectory(src/util)
+  target_link_libraries(${PROJECT_NAME} util)
 
   # My libraries: snail
   add_subdirectory(src/snail)
