+++ android/external/Android.mk
@@ -1,3 +1,3 @@
 include $(call all-subdir-makefiles)
 LOCAL_LDLIBS := -L$(SYSROOT)/usr/lib -llog
-LOCAL_SHARED_LIBRARIES := lib snail elona
+LOCAL_SHARED_LIBRARIES := util snail elona
