--- android/external/lib
@@ -1,1 +1,0 @@
-../../src/lib
