+++ src/elona/lua_env/handle_manager.hpp
@@ -5,7 +5,7 @@
 #include <boost/uuid/uuid.hpp>
 #include <boost/uuid/uuid_generators.hpp>
 #include <boost/uuid/uuid_io.hpp>
-#include "../../lib/noncopyable.hpp"
+#include "../../util/noncopyable.hpp"
 #include "../item.hpp"
 #include "lua_env.hpp"
