+++ src/elona/message.hpp
@@ -3,8 +3,8 @@
 #include <string>
 #include <utility>
 #include <vector>
-#include "../lib/namedparam.hpp"
 #include "../snail/color.hpp"
+#include "../util/namedparam.hpp"
