--- src/lib/CMakeLists.txt
@@ -1,8 +1,0 @@
-# Source files
-set(LIB_SOURCES
-  fps_counter.cpp
-  profiler.cpp
-  )
-
-add_library(lib STATIC ${LIB_SOURCES})
-
