+++ src/snail/scene.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <stack>
-#include "../lib/noncopyable.hpp"
+#include "../util/noncopyable.hpp"
 #include "node.hpp"
