+++ src/snail/window.hpp
@@ -2,8 +2,8 @@
 
 #include <memory>
 #include <string>
-#include "../lib/enumutil.hpp"
-#include "../lib/noncopyable.hpp"
+#include "../util/enumutil.hpp"
+#include "../util/noncopyable.hpp"
 #include "detail/sdl.hpp"
