+++ src/util/Android.mk
@@ -2,7 +2,7 @@ LOCAL_PATH := $(call my-dir)
 
 include $(CLEAR_VARS)
 
-LOCAL_MODULE := lib
+LOCAL_MODULE := util
 
 SDL_PATH := ../../android/external
