+++ src/util/CMakeLists.txt
@@ -1,0 +1,8 @@
+# Source files
+set(UTIL_SOURCES
+  fps_counter.cpp
+  profiler.cpp
+  )
+
+add_library(util STATIC ${UTIL_SOURCES})
+
