+++ src/util/enumutil.hpp
