+++ src/util/fps_counter.cpp
